Update to use a single DateTimePicker
