--- src/styles/_date-picker.scss
@@ -1,16 +1,0 @@
-@use './typography';
-@use './colors';
-
-@mixin all {
-  .seam-date-picker {
-    position: relative;
-
-    > .seam-native-picker {
-      position: absolute;
-      visibility: hidden;
-      bottom: 0;
-      left: 0;
-      width: 100%;
-    }
-  }
-}
