--- src/styles/_time-picker.scss
@@ -1,16 +1,0 @@
-@use './typography';
-@use './colors';
-
-@mixin all {
-  .seam-time-picker {
-    position: relative;
-
-    > .seam-native-picker {
-      position: absolute;
-      visibility: hidden;
-      bottom: 0;
-      left: 0;
-      width: 100%;
-    }
-  }
-}
